Gather all options in a single file
