--- torchqdynamics/mesolve/options.py
@@ -1,18 +1,0 @@
-from ..options import ODEFixedStep
-
-all = ['Rouchon', 'Rouchon1', 'Rouchon1_5', 'Rouchon2']
-
-
-class Rouchon1(ODEFixedStep):
-    pass
-
-
-Rouchon = Rouchon1
-
-
-class Rouchon1_5(ODEFixedStep):
-    pass
-
-
-class Rouchon2(ODEFixedStep):
-    pass
