--- torchqdynamics/sesolve/options.py
@@ -1,7 +1,0 @@
-from ..options import Options
-
-all = ['Propagator']
-
-
-class Propagator(Options):
-    pass
